refactor(status): extract collectStatus into shared read-only src/status module
